+++ Adjust/sdk-core/src/main/java/com/adjust/sdk/Util.java
@@ -637,4 +637,8 @@ public class Util {
 
         return false;
     }
+
+    public static String getSdkVersion() {
+        return Constants.CLIENT_SDK;
+    }
 }